Move Side.h to controllers and rename to side.h
The move makes the dependency direction only robotInterfaces deps
on controllers, instead of circular.  The rename just comes for
free since I'm here anyway.
