+++ drake/examples/humanoid_controller/BUILD.bazel
@@ -44,8 +44,8 @@ drake_cc_library(
         "//drake/common:essential",
         "//drake/multibody:rigid_body_tree",
         "//drake/multibody:rigid_body_tree_alias_groups",
+        "//drake/systems/controllers:side",
         "//drake/systems/controllers/qp_inverse_dynamics:robot_kinematic_state",  # noqa
-        "//drake/systems/robotInterfaces:side",
     ],
 )
