+++ drake/systems/robotInterfaces/QPLocomotionPlan.h
@@ -13,9 +13,9 @@
 #include "drake/common/unused.h"
 #include "drake/lcmt_qp_controller_input.hpp"
 #include "drake/multibody/rigid_body_tree.h"
+#include "drake/systems/controllers/side.h"
 #include "drake/systems/controllers/zmpUtil.h"
 #include "drake/systems/robotInterfaces/BodyMotionData.h"
-#include "drake/systems/robotInterfaces/Side.h"
 
 class QuadraticLyapunovFunction {
   // TODO(tkoolen): move into its own file
