+++ tools/install/libdrake/build_components.bzl
@@ -232,6 +232,7 @@ LIBDRAKE_COMPONENTS = [
     "//drake/systems/controllers:pid_controller",
     "//drake/systems/controllers:qp_common",
     "//drake/systems/controllers:setpoint",
+    "//drake/systems/controllers:side",
     "//drake/systems/controllers:state_feedback_controller_interface",
     "//drake/systems/controllers:zmp_planner",
     "//drake/systems/controllers:zmp_util",
@@ -304,7 +305,6 @@ LIBDRAKE_COMPONENTS = [
     "//drake/systems/rendering:pose_vector",
     "//drake/systems/robotInterfaces:body_motion_data",
     "//drake/systems/robotInterfaces:qp_locomotion_plan",
-    "//drake/systems/robotInterfaces:side",
     "//drake/systems/sensors:accelerometer",
     "//drake/systems/sensors:beam_model",
     "//drake/systems/sensors:beam_model_params",
